backend/pyproject.toml
[project]
name = "altra-interventions"
version = "0.1.0"
dependencies = [
"fastapi==0.115.0",
"uvicorn[standard]==0.30.6",
"psycopg2-binary==2.9.9",
"SQLAlchemy==2.0.35",
"pydantic==2.9.1",
"httpx==0.27.2",
"python-dotenv==1.0.1",
"python-multipart==0.0.9",
"pyjwt[crypto]==2.9.0",
]


[tool.uvicorn]
factory = true
host = "0.0.0.0"
port = 8000